Refactoring metodos Punto Transportify
